refactor StandardTable - extract columns - support multiple sums display
